requirements.txt
dbt-oracle==1.5.0
pandas==2.1.3
oracledb==1.3.0
openpyxl==3.11.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
